staging_node_cli
Module chain_spec

Substrate chain configurations.

Structs

Extensions

Node ChainSpec extensions.

RuntimeGenesisConfig

Functions

development_config

Development config (single validator Alice).

flaming_fir_config

Flaming Fir testnet generator

local_testnet_config

Local testnet config (multivalidator Alice + Bob).

staging_testnet_config

Staging testnet config.

testnet_genesis

Helper function to create RuntimeGenesisConfig json patch for testing.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Balance

Balance of an account.

ChainSpec

Specialized ChainSpec.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.